Create README.md
A complete Java web application. Which is implemented to perform the online examination for the student. Where an admin can add questions and Types of examination depends on the category.
User Login: To access the exam user must be registered in the application. Once the user will register in the examination system and project user will be able to view all exams. The user can choose an exam and start the test.

Admin Login: Admin can log in to the application and perform operations like manage exams, manage users, check the results, manage questions, and download the result pdf.